% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{assert_no_special_chars}
\alias{assert_no_special_chars}
\title{Raise an error if any of the args contain special characters}
\usage{
assert_no_special_chars(args)
}
\arguments{
\item{args}{character vector of args to check}
}
\description{
Raise an error if any of the args contain special characters
}
