% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{assert_sane_character_vector}
\alias{assert_sane_character_vector}
\title{Assert character vector is sane.}
\usage{
assert_sane_character_vector(args)
}
\arguments{
\item{args}{Character vector of arguments to check.}
}
\value{
Logical value.  TRUE if all tests pass, otherwise throw an error.
}
\description{
Zero-length vector allowed. No NAs allowed. Must be fewer than 1000 arguments
and fewer than 200000 characters.
}
