% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{escape_spaces}
\alias{escape_spaces}
\title{Escape any plain spaces.}
\usage{
escape_spaces(args)
}
\arguments{
\item{args}{A sane character vector of arguments. See \code{assert_sane_character_vector()}}
}
\value{
Character vector with spaces escaped with a backslash.
}
\description{
Usually only apply this to arguments which represent filenames and contain
filename expansion characters.
}
