% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{has_file_expansion}
\alias{has_file_expansion}
\title{Detect special characters in a character vector of args}
\usage{
has_file_expansion(args)
}
\arguments{
\item{args}{character vector of args to check}
}
\value{
logical vector to match input 'args' set to TRUE if there are file
        expansion characters in the individual arguments.
}
\description{
Detect special characters in a character vector of args
}
