% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{has_spaces}
\alias{has_spaces}
\title{Detect plain spaces in a character vector.}
\usage{
has_spaces(args)
}
\arguments{
\item{args}{Character vector of arguments to check.}
}
\value{
>ogical vector. Elements are set to TRUE if there are spaces
        in the corresponding \code{args} element.
}
\description{
This function checks for plain ASCII space characters. It
ignores tab, newline and any other "whitespace" characters.
}
