% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run the given command the supplied arguments. Shell expansion of filenames is still performed.}
\usage{
run(command, args = NULL, error_on_status = TRUE, echo = FALSE, timeout = 10)
}
\arguments{
\item{command}{Character scalar, the command to run}

\item{args}{character vector, arguments to the command}

\item{error_on_status}{Throw an error if the command returns a non-zero status?
Default: TRUE}

\item{echo}{Echo the command to be run. Default: TRUE}

\item{timeout}{whole seconds for which to run before interrupting process.
Default: 0 (no limit)}
}
\value{
list with 'status', 'stdout', 'stderr'
}
\description{
Run the given command the supplied arguments. Shell expansion of filenames is still performed.
}
