% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{shell_quote}
\alias{shell_quote}
\title{Shell quote all arguments except for those which include bash filename expansion}
\usage{
shell_quote(args)
}
\arguments{
\item{args}{sane character vector of args. See \code{assert_sane_character_vector()}}
}
\value{
Apply shQuote() to all arguments which do not look like they do
        not contain base filename expansion characters
}
\description{
Filename expansion characters are \code{*}, \code{[} and \code{?}
}
