% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{assert_sane_character_vector}
\alias{assert_sane_character_vector}
\title{Assert character vector is sane}
\usage{
assert_sane_character_vector(args)
}
\arguments{
\item{args}{character vector of args to check}
}
\description{
Zero-length vector allowed.
No NAs allowed.
}
