% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{escape_spaces}
\alias{escape_spaces}
\title{Escape any spaces in args e.g. in filenames}
\usage{
escape_spaces(args)
}
\arguments{
\item{args}{sane character vector of args. See \code{assert_sane_character_vector()}}
}
\value{
character vector with spaces escaped with a backslash.
}
\description{
Escape any spaces in args e.g. in filenames
}
