% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{has_spaces}
\alias{has_spaces}
\title{Detect plain spaces in the args}
\usage{
has_spaces(args)
}
\arguments{
\item{args}{character vector of args to check}
}
\value{
logical vector to match input 'args' set to TRUE if there are spaces
        in the individual arguments.
}
\description{
Detect plain spaces in the args
}
