% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{has_special_chars}
\alias{has_special_chars}
\title{Detect special characters in a character vector of args}
\usage{
has_special_chars(args)
}
\arguments{
\item{args}{character vector of args to check}
}
\value{
logical vector to match input 'args' set to TRUE if there are special
        characters in the individual arguments.
}
\description{
This is implemented as a whitelist of characters to accept. The presence
of anything outside this whitelist is considered a 'special character'
}
